Gibber rebalance (#33737)
:cl: coiax
balance: The kitchen gibber must be anchored in order to use.
balance: The gibber requires bodies to have no external items or
equipment.
/:cl:

Unanchoring the gibber, running around with it, ebowing people, then
shoving them in is not fun.

As a double check, you REALLY should be stripping them of all that
cloth, it'll jam up the gears.

- Removes the abiotic() proc, it's literally used only in the gibber.