Reduces the memory footprint of the gas_mixture datum
To save memory, byond only stores vars that are written to, referring to the original compile time value stored once for each type, otherwise.

Originally we set defaults in New() to control the order of the vars in this changed vars list to optimize access speed, but this no longer makes sense to do because byond moved that from a linked list to a binary insertion sorted vector.

And making the volume default to the most popular case, and only get written too otherwise should be a noticeable memory improvement.