Reduce the volume of the death sound on clock work mobs (#34452)
* Reduce the volume of the death sound on clock work mobs

I edited the .oog since I did not see a way in the code to lower the volume. This will affect all code/game/gamemodes/clock_cult/clock_mobs which at this moment is only maraders and the eminence.

* Non-clockwork mobs are no longer affected

adjusted volume in the code so that the clockwork golems, ratvar, and ratvar plushies are unaffected by the audio file changes