One can no longer use cogscarab as a form of Eminence priority. (#34611)
Cogscarabs are meant to build, and wasting an integral (building) slot for something you can get as a ghost so you get priority is shitty to both other ghosts and team (the second one being an admin issue).

If you want to go eminence, you can roll for it as a ghost like everyone else.