Mamma-mia! Adds-a the Italian-a language! (#34916)
* Mamma-mia!

That's a one-a spicy meatball! The first of many spicy meatballs!

* I have-a returned from the exile and fixed-a the things people asked-a for!

Includes changing sec so that not every word with sec in it will end
with problems. Uses resident italian advice to make minor edits. The
chef is now mario-tier of not-italian.

* Adds-a the mutation code and-a calls it for the cook at roundstart

Mamma-mia whoever named-a the CLOWNMUT and-a made it a #define trolled-a
the hell out of me-a! It was like-a chasing an actual-a clown!

* Actually bothers to upload the trait definition

Lesson gained: Save all, all the time.

* Case sensitive fun whee

Someday I'll learn to build first

* IT COMPILES-A

Fucking finally.

* Small addition that was left out and attempting to merge resolve the odd DME change

I never actually edited the dme in any way, only dm files; I think I
just haven't updated the master recently, so the exe is in a different
place than the repo. Don't expect this to compile.

* Fixes .dme to match with master

If this doesn't work I'll start reverting to the last version that didn't throw a fit.

* Commit to get travis to recheck

.dmm should be fixed, so this needs rechecking. Also, adding wine
replacement.

* The pizza, she is a-ready to share-a with the people!

Fixes a lot, does as much sanitization as is possible so that words like
insect don't become inpoliziact or similar.  For some words, punctuation
may be displaced by the final space necessary to keep normal words from
running into eachother.  This part, at least, is done.