Newliquors (#35581)
* Adds a few new liqueurs and cocktails

How the fuck in THIS game of all games do we not have Triple Sec.
SERIOUSLY.

* Well there goes a one commit clean PR rip me

reeeeee

* Adds sounds, fixes colors

* Fixes recipies giving eggnog

Don't ask

* Adds space where space should be

* Travis stop stalling reeeeeee

* Fixes spelling mistakes

* Adds some health benefits to sec drinks.

Quadruple sec is mild, but easy to make; quintuple sec is STRONG but
more difficult.

* Fixes merge conflict annoyance

* FUCK

* Updates boozepwr of Quadruple/Quintuple sec

If someone ever adds sex on the beach as a cocktail, I fully expect them
to add in Sex-tuple Sec as well.

* Makes clowntears reagent explicit within food_reagents

Removed hacky solution, introduced better one