Adds Medical Sprays (#36081)
* Adds Medical Sprays

Medsprays are a new application method for chemicals, functioning like a multi-use patch with an internal chemical reservoir.

This allows touch chemicals to be applied more precisely and with a delay, rather than instantly and imprecisely with a spray bottle.

Pre-loaded medsprays with styptic powder or silver sulfadiazine can be obtaiend from the NanoMeds, and empty ones will spawn in the secure chemical locker.

They will also come from the medical pack from cargo.

* Update packs.dm

* Volume adjustment, attack_self addition

* Adds Medical Sprays

Medsprays are a new application method for chemicals, functioning like a multi-use patch with an internal chemical reservoir.

This allows touch chemicals to be applied more precisely and with a delay, rather than instantly and imprecisely with a spray bottle.

Pre-loaded medsprays with styptic powder or silver sulfadiazine can be obtaiend from the NanoMeds, and empty ones will spawn in the secure chemical locker.

They will also come from the medical pack from cargo.

* Volume adjustment, attack_self addition

* Adds Medical Sprays

Medsprays are a new application method for chemicals, functioning like a multi-use patch with an internal chemical reservoir.

This allows touch chemicals to be applied more precisely and with a delay, rather than instantly and imprecisely with a spray bottle.

Pre-loaded medsprays with styptic powder or silver sulfadiazine can be obtaiend from the NanoMeds, and empty ones will spawn in the secure chemical locker.

They will also come from the medical pack from cargo.

* Volume adjustment, attack_self addition

* Forgot this

* removes unnecessary/annoying check

* replaces deprecated sterilizer spray bottles on Pubby

* Applies MrDoomBringer's sprites

* new sprites, by @MrDoomBringer

* two-line shuffle

* fixes double entry

* Adds logging

* code standards