Adds a French beret! HONH HONH HONH! (#37457)
This is very similar to the Italian Moustache, but it's a beret. This will be part of a series of language items. Happy to entertain suggestions as to whether it should be a roundstart item or randomly scattered about maint/cargo crate restricted. In the spirit of atomizing PRs, this one will just be for the French Beret.

Ready for merge.