Replaces adding goliath plates to clothing and APLUs with a customizable tracking component (#37525)
add: Amount of goliath plates is now shown in examine for explorer suits and mining hardsuits.


I was peeved by being unable to tell whether the armour is upgraded or not except by trying to stick more plates on it.

I think a more correct way would be to add a plate amount var to these items similarly to APLU plating and then have description mention their amount and, for added transparency, that it's possible to improve it if there's 0 plates.

Please point the errors/wrongs in it, since "current armour value minus constant divided by value each plate adds" is rather clunky code.