Adds Fernet, a digestif that reduces your satiety. (#37869)
* Added Fernet and subdrinks

Added sprites for Fernet (bottle), and glasses of pure Fernet, Fernet Cola, Fanciulli and Branca Menta

* Adds Fernet and drinks made out of it

Added Fernet, Fernet Cola, Fanciulli and Branca Menta

* Adds reactions to fernet

* Adds Bottle of Fernet to the list

Spriting by the wonderful NeoExperiences

* Adds Fernet to the Borg's Shaker

They can help you indulge in more hedonism now.

* Adds Fernet to the Booze Dispenser

* Adds Bottle of Fernet to the list

* Updated Fanciulli, Branca Menta

Oops, nearly made two of the drinks not cause the satiety reduction.

* Fixed a missing comma on reagent list

I'm dumb, I forgot this was a list

* Fixed the procs on stamina and hunger

I'm sorry I'm sleepy and retarded to mistake an L for an M

* Changes the recipe of Fanciulli to include Manhattan

They were conflicting, and the description of the glass already says it's just Manhattan with Fernet, so it's consistent.

* Removes glass_icon_state from pure fernet

Cobby said it was pointless since the spritework was lazy