Improves descriptions of Integrated Circuits (#37678)
* Fixes a ton of typos, inaccuracies, poor grammar, and confusing weirdnesses from the descriptions of almost every integrated circuit.
For example, the reagent filter:

Before:
"This is a filter which will move liquids from the source to the target. It will move all reagents, except those in the unwanted list, given the fourth pin if amount value is positive, or it will move only desired reagents if amount is negative. The third pin determines how much reagent is moved per pulse, between 0 and 50. Amount is given for each separate reagent."

After:
This is a filter which will move liquids from the source to its target.
If the amount in the fourth pin is positive, it will move all reagents except those in the unwanted list.
If the amount in the fourth pin is negative, it will only move the reagents in the wanted list.
The third pin determines how many reagents are moved per pulse, between 0 and 50. Amount is given for each separate reagent."

* fixed, ready

* Actually fixed and ready.

* Reverts the name of the coordinate pathfinder circuit back to "coordinte"

* Fixes a few missed, or poorly rewritten parts.

* Update smart.dm