Removes normal ways to obtain drones (#38101)
I don't want this merged as soon as, I'd just rather normal drones to go away if no-one offers to fix their broken state.

Going to see if I can find out from someone how to edit maps tomorrow to remove the dispensers from them since you can't make them and I swear I've seen them before.

    remove any way to obtain drones outside the dispenser
    remove dispenser from maps

Edit: To better explain for the sake of a good PR body and a lack of a link to the PR a day ago, drones have been in the code in roughly the same state for 3 years (dispenser PR by Xhuis is missing from GitHub) and have gained laws admins enforce when they notice drones doing the wrong thing due to it being a ghost role. The code verbally dictates this will happen. The role hasn't been coded to prevent these many wrong things and has weak balance to begin with being a monkey with a vent ready backpack.

Administration hasn't stopped caring about this role. It should be in their direct hands until the role is clearly not an administration problem and fixed to represent that.