PDA can be ctrl-clicked to remove the item in its pen slot (#38577)
* Makes pens ez to remove from pda

Players can now ctrl click on the PDA to remove the pen without having to remove the ID first.

* added instructions when examined.

Like how it tells the user they can use alt-click to remove contents, it now says they can use ctrl-click to remove items from the pen slot.

* PDA can now be dragged without dropping the pen.

When ctrl-clicked it now checks to see if the PDA is on a turf and will only output the pen if its not.

* quick change to examine text.

* comma

* Improved the examine