Admin spawnable version of the Death Ripley that actually inflicts damage and can rip arms off (#38663)
* On Harm intent causes damage

* Doesnt push people away when attacking them in harm intent

* removed emtpy line

* Revert "removed emtpy line"

This reverts commit 992bd05631f4d8d1b6a6aa851cd35137c4e94c73.

* Now rips both arms off

Unless only the left arm is missing, in which case nothing happens and im not sure why. If the right arm is missing though the left arm is still ripped off.

* Fixes issues with null values and such

added if statements that check if the affected limbs are null. Also added the attack to the logs.

* Makes a "proper" version of the death clamp

Creates a "proper" version of the death clamp that causes damage. Also added a "proper" version of the Death Ripley that includes said death clamp.

* ShizCalev's change requests

-Uses TRUE instead of 1.
-Used less stupid naming conventions and made the objects improper nouns (why did I even make them proper?)
-Gives a detailed log of which limbs were dismembered

* Cyberboss's requested changes

-Moved variable declaration
-Playsound changed
-Added equipment.Cut() to initialization