Allows empty borg shells to be disassembled. (#38639)
* Adds the ability to deconstruct empty cyborg shells.

* Update robot_parts.dm

* Able to directly swap cells

Added ability to change out power cells with a screwdriver.

* Update robot_parts.dm

* Removing unnecessary `src` prefixes from code.

* Update robot_parts.dm

* Moved to wrench_act and screwdriver_act overrides

I think I did this right

* Replaced screwdriver code

* Swapping out qdeleted for an if-not

* Adds a check for cell-less shells when MMI is added.

* Update robot_parts.dm

* Adds the ability to just remove the cell

Re-organized the screwdriver section as well, and added the ability to add a cell to a shell with no cell.

* Whoops

Copypaste error

* Changed the screwdriver section from nested if statements to use a function

Also removed the line that forced the naked endoskeleton out of your hand in the wrench section.