gives it 7 spread!
Why not give the ops a 300 TC instantly obtain the disk button? Or a 300 tc mech that can't be damaged in any way except by pulse lasers?

Just because something is high commitment doesn't mean you can put it at a power level that the station cannot fucking deal with.

someone actually holds this opinion