Added max and min circuit (#39013)
* Added max and min circuit

This now has a max and min circuit which checks all values and returns the highest/lowest of them. Tested and working.

* update: saving lines of code

The more compact min max. you may as well want to try the same with addition and substraction.

* Reformated code and improved proc

As you said, added early continue, put the vars low and changed the logic on the result.

* Update arithmetic.dm