Changes consoles on the old box white ship ruin into "broken computers" (#39176)
We get a lot of issue reports and admin PMs telling us that the ship doesn't
work when it was never supposed to fly. Then why have two consoles that give
the illusion it does, saying "shuttle: missing" like it's bugged?

This removes that ambiguity by turning them into the same green broken
computers as on the ancient station space ruin.