Increase Odysseus mech movespeed (#39357)
Right now, the Odysseus is in an awkward spot where it is worse at transporting
critical patients than a player with a stabilized light pink slime extract, a
medical cyborg or even just a player with an epipen+roller bed/office chair.
This is due to its low movespeed - since it moves as fast as a Gygax, you'll
get the patient to medbay quicker if you just get out and drag them.

Setting step_in to 2 will make it as fast as a Phazon, which imo won't result
in balance issues if you consider that it has less max integrity than a HONK
and its only weapon is completely blocked by thick clothing.