Fix the cig reagent bug, remove pointless and buggy RNG (#39334)
The balance in the note is no longer a concern, a 50/50 sacid and facid dipped
cig took almost a minute to knockout and damage stopped when it was removed. It
allowed a severe bug in the chem system.

I'm not closing that issue report because in theory the underlying bug is still
present and any future ways to apply chems may still have it.