Refactor and improve briefcase launchpad (#39586)
According to statbus, the bluespace launchpad is purchased in less than 2% of
rounds on average. It's obvious why - it's slow, loud and extremely noticeable
with a short range.

I refactored it to storage/briefcase/launchpad (thanks XDTM) and disguised the
remote as a folder that spawns inside the briefcase. This will make it harder
to detect than a briefcase that can't actually hold items and a red remote that
screams "hi I'm valid". On top of that, I increased the teleport range from 3
to 8 (roughly half the screen). It still has the slow and very obvious
animation, but can finally be used to retrieve and send items that aren't at
the corner of a room.