Make the inventory helper overlay less impactful (#39824)
Reduces alpha and has a traffic light system for the items

-> red ghost when the item can't be placed
-> green ghost when it can
-> full solid colour item, object is in the inventory

This is so you can differentiate the ghost view from the actual object entering
the slot, which was sticky's primary complaint

I moved the cut of overlays only to mouse out, since if that somehow isn't
called we have bigger problems than overlays doubling up. INB4 this causes
server crashes.

Side bonus it removes one image operation, it's worth investigating a cache for
this as well I would say, a simple rotating buffer list of the last x many
images.

Tested and ready to go