IC GPS Update (#39630)
GPS circuit now has a 4th output, placing X,Y,Z all in a string.
2 new converters: Rel to Abs, and Adv Rel to Abs.
Rel to Abs takes a set of relative and a set of absolute coordinates, and converts the relative one to absolute. 1 complexity.
Adv Rel to Abs takes a set of relative coordinates and converts it to absolute without the need for an already known set of absolute coordinates. 2 complexity.