oof ouch my bones (#40005)
I now have a more finalized version of it.
The recipe is 3 itching powder, 1 unstable mutagen, 1 milk
On normal usage, it does 15 stamina damage a cycle and consumes at a rate of 2 units per cycle. If you're a skellyton or a plasmaman it does 0.5 brute as well.
On overdose, it has a small chance of causing a huge spike of stamina damage to a random limb, temporarily disabling it. If you're a skellyton or a plasmaman, it does the same thing but with brute instead of stamina.
There's some other effects as well such as some forced speech but those are all not really gameplay affecting.