mannequin fix for lizard people (#40308)
* mannequin fix for lizard people

Just a small sequence problem with the race default features being loaded after setting them and tail not being set to update.

* Mannequin Fix for Lizard People

Just a small sequence problem with the race default features being loaded after setting them and tail not being set to update.

This would lead to the mannequin just having a smooth tail and no spines the default setting for a lizard tail.