Allows lethal turrets to target blobs (#41292)
cl ShizCalev
balance: Lethal turrets can now target blobs.
/cl

Gives an AI a way to defend itself if a blob decides to plop itself down in it's core or the AI upload.

Likewise, this also allows the crew to build laser gun turrets to help fight against it, providing a little more challenge to the blob if the crew gets smart.

Downside is that turrets are still very weak again blob (they die in a single hit by a blob), so it's not a be-all and end-all defensive strat.